feat: refactor Doctor model and related services; update relationships to use userId as primary key and add Notifications model

+++ src/doctor/doctor.service.ts
@@ -42,7 +42,7 @@ export class DoctorService {
         }
     }
 
-    async getAllDoctors(page: number, limit: number, specialization: string, experience: number[], weeks: string[], fees: number[], isActive: boolean, search: string) {
+    async getAllDoctors(page: number, limit: number, specialization: string, experience: number[], weeks: string[], fees: number[], search: string) {
 
         const query: any = specialization ? { specialization: { contains: specialization, mode: 'insensitive' as const } } : {} // will filter case-insensitive
 
@@ -51,24 +51,15 @@ export class DoctorService {
             query['experience'] = { gte: min, lte: max };
         }
 
-        if (weeks?.length) {
-            query['OR'] = weeks.map(week => ({
-                availableTimes: {
-                    some: { contains: week }
-                }
-            }))
-        }
-
         if (fees?.length === 2) {
             const [min, max] = fees;
             query['fees'] = { gte: min, lte: max };
         }
 
-        if (isActive) query['isActive'] = isActive
+        // if (isActive) query['isActive'] = isActive
 
         if (search) {
             query.OR = [
-                ...(query.OR || []), // preserve existing OR filters (like weeks)
                 { specialization: { contains: search, mode: 'insensitive' } },
                 { education: { contains: search, mode: 'insensitive' } },
                 { aboutMe: { contains: search, mode: 'insensitive' } },
@@ -126,7 +117,7 @@ export class DoctorService {
 
         try {
             const fetchedDoctor = await this.prisma.doctor.findUnique({
-                where: { id },
+                where: { userId: id },
                 select: doctorSelect
             })
 
@@ -145,7 +136,7 @@ export class DoctorService {
                     where: {
                         specialization: fetchedDoctor?.specialization,
                         isActive: true,
-                        id: {
+                        userId: {
                             not: id,
                         }
                     },
@@ -155,19 +146,19 @@ export class DoctorService {
             ])
 
             //sort doctors based on average rating
-            const sortedRelatedDoctors = this.modifyDoctors(relatedDoctors)
+            // const sortedRelatedDoctors = this.modifyDoctors(relatedDoctors)
 
-            // pagination reviews
-            const totalItems = fetchedDoctor?.reviews.length
-            let totalPages: number = 0
+            // // pagination reviews
+            // const totalItems = fetchedDoctor?.reviews.length
+            // let totalPages: number = 0
 
-            if (totalItems) {
-                totalPages = Math.ceil(totalItems as number / limit)
+            // if (totalItems) {
+            //     totalPages = Math.ceil(totalItems as number / limit)
 
-                const paginatedReviews = fetchedDoctor?.reviews.slice(skip, skip + limit)
+            //     const paginatedReviews = fetchedDoctor?.reviews.slice(skip, skip + limit)
 
-                fetchedDoctor.reviews = paginatedReviews
-            }
+            //     fetchedDoctor.reviews = paginatedReviews
+            // }
 
             return {
                 data: {
@@ -176,13 +167,13 @@ export class DoctorService {
                         averageRating: averageRating._avg.rating,
                         totalReviews
                     },
-                    relatedDoctors: sortedRelatedDoctors,
-                    pagination: {
-                        totalItems,
-                        totalPages,
-                        currentPage: page,
-                        itemsPerPage: limit
-                    }
+                    // relatedDoctors: sortedRelatedDoctors,
+                    // pagination: {
+                    //     totalItems,
+                    //     totalPages,
+                    //     currentPage: page,
+                    //     itemsPerPage: limit
+                    // }
                 },
                 message: "Doctor fetched successfully"
             }
@@ -233,7 +224,7 @@ export class DoctorService {
         try {
 
             const doctor = await this.prisma.doctor.update({
-                where: { id },
+                where: { userId: id },
                 data: { specialization, education, experience, aboutMe, fees, availableTimes }
             })
 
@@ -257,7 +248,7 @@ export class DoctorService {
         try {
 
             const doctor = await this.prisma.doctor.update({
-                where: { id },
+                where: { userId: id },
                 data: {
                     availableTimes: {
                         push: availableTime
@@ -283,14 +274,14 @@ export class DoctorService {
     async removeAvailableTime(id: string, availableTime: string) {
 
         try {
-            const doctorRecord = await this.prisma.doctor.findUnique({ where: { id } });
+            const doctorRecord = await this.prisma.doctor.findUnique({ where: { userId: id } });
 
             if (!doctorRecord) this.handleErrorsService.throwNotFoundError("Doctor not found")
 
             const updatedAvailableTimes = doctorRecord?.availableTimes.filter((time: string) => time !== availableTime);
 
             const doctor = await this.prisma.doctor.update({
-                where: { id },
+                where: { userId: id },
                 data: {
                     availableTimes: updatedAvailableTimes
                 }
@@ -313,7 +304,7 @@ export class DoctorService {
         try {
 
             const doctor = await this.prisma.doctor.update({
-                where: { id },
+                where: { userId: id },
                 data: {
                     isActive: true
                 }
@@ -336,7 +327,7 @@ export class DoctorService {
         try {
 
             const doctor = await this.prisma.doctor.update({
-                where: { id },
+                where: { userId: id },
                 data: {
                     isActive: false
                 }
@@ -357,7 +348,7 @@ export class DoctorService {
     async deleteDoctor(id: string) {
 
         try {
-            const doctor = await this.prisma.doctor.delete({ where: { id } })
+            const doctor = await this.prisma.doctor.delete({ where: { userId: id } })
 
             if (!doctor) this.handleErrorsService.throwNotFoundError("Doctor not found")
